requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
chromadb
pypdf
streamlit
python-dotenv
sentence-transformers
pysqlite3-binary
ddgs
duckduckgo-search
langchain-openai
youtube-transcript-api
pytube